#49 - moves date picker field to core ui module
