+++ core/ui/src/androidTest/java/br/com/rbrthmn/ui/components/DatePickerFieldUITest.kt
@@ -1,24 +1,4 @@
-/*
- *
- * Copyright (C) 2022 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Modifications made by Roberto Kenzo Hamano, 2024
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package br.com.rbrthmn.operations
+package br.com.rbrthmn.ui.components
 
 import androidx.activity.ComponentActivity
 import androidx.compose.ui.test.assertIsDisplayed
@@ -27,9 +7,8 @@ import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
-import br.com.rbrthmn.operations.ui.components.CALENDAR_TEST_TAG
-import br.com.rbrthmn.operations.ui.components.DatePickerField
 import br.com.rbrthmn.ui.BaseUITest
+import br.com.rbrthmn.ui.R
 import br.com.rbrthmn.ui.onNodeWithStringId
 import org.junit.Test
